requirements.txt
# API
fastapi
uvicorn[standard]
python-multipart

# UI
streamlit
requests

# Core ML/Vector Search
langchain
langchain-community
sentence-transformers
faiss-cpu
PyMuPDF

# New dependencies
rank-bm25
python-docx
httpx

# Utils
python-dotenv

# Testing
pytest
pytest-asyncio
